requirements.txt
customtkinter==5.2.2
Pillow==12.0.0
